Refactor kick decisions to a kick_selector
