+++ crates/control/src/lib.rs
@@ -11,6 +11,7 @@ pub mod game_controller_filter;
 pub mod game_state_filter;
 pub mod ground_contact_detector;
 pub mod ground_provider;
+pub mod kick_selector;
 pub mod kinematics_provider;
 pub mod led_status;
 pub mod limb_projector;
